PET-92 refactor : Test-Module 파일 testFixture 패키지로 이동
