+++ Common-Module/src/testFixtures/java/com/pawith/commonmodule/BaseRestDocsTest.java
@@ -1,4 +1,4 @@
-package com.pawith.testmodule;
+package com.pawith.commonmodule;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -13,7 +13,6 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
-import org.springframework.web.filter.CharacterEncodingFilter;
 
 @AutoConfigureRestDocs
 @Import({RestDocsConfig.class})
@@ -32,7 +31,7 @@ public class BaseRestDocsTest {
             .apply(MockMvcRestDocumentation.documentationConfiguration(provider).uris().withPort(8080))
             .alwaysDo(MockMvcResultHandlers.print())
             .alwaysDo(resultHandler)
-            .addFilters(new CharacterEncodingFilter("UTF-8", true))
+//            .addFilters(new CharacterEncodingFilter("UTF-8", true))
             .build();
     }
